pi_mutex: Make error codes consistent with glibc counterparts
The current implementation for pi_mutex* functions returns 0 on
success and a negative error number to indicate the error.

POSIX is unclear on the sign of the error number specifying only that
"functions shall return zero; otherwise, an error number shall be
returned to indicate the error".

The glibc "de facto" standard for pthread_mutex* is to return 0 on
success and a positive error code on failure (e.g. EINVAL instead of
-EINVAL for an invalid parameter).

While this implementation is non-POSIX it is still beneficial to match
the glibc behavior in order to avoid porting/programming mistakes.
Change return paths in librtpi to match the glibc behavior i.e.
return a positive error number on failure.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>