pi_mutex: Clear stored tid for child thread after fork
The current implementation of gettid() stores the current thread id in
a thread local static variable 'tid_this_thread' for optimization
purposes i.e. it avoids a SYS_gettid syscall for every access.

This creates a problem for a child thread after a fork. When fork() is
called in a multithreaded process, only the calling thread is
duplicated in the child process. This child thread will have a
different thread id however the old value continues to be returned
since 'tid_this_thread' is non-zero.

Install an 'atfork' handler for the child process which clears the
stored tid for the duplicated thread.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>