glibc-tests: Skip tst-cond12 and by extension tst-cond13
The current librtpi API creates a strong association between a user
mutex and a conditional variable by passing a pointer to the mutex to
pi_cond_init() along with the pointer to the conditional variable to
be initialized. The pointer to the associated mutex is stored in the
private data for the conditional variable.

This has several advantages like preventing the practice of signaling
the condition prior to the association of the mutex, or allowing the
use of the user mutex in the signaling, destroy, and cancellation
paths to protect the conditional variable private data accesses w/o
requiring an internal private lock which can be an additional source
of races/errors.

There is however one corner case where this API breaks down when it
comes to process shared conditional variables, exposed by tst-cond12
and tst-cond13 (which re-uses the tst-cond12 code). In these tests a
parent process creates a memory mapping for a process shared
conditional variable and mutex that is later re-mapped by the child
process at a different address and the old mapping is unmapped via
munmap(). Because the conditional variable was initialized before the
fork() call the stored mutex pointer now points to an unmapped address
in the child process and it results in a SIGSEGV crash when the mutex
is accessed inside the pi_cond code.

Skip these tests for now until we decide if this use case is worth the
API redesign required to fix this issue.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>