pi_mutex: Add mutex lock save/restore helper macros
In certain scenarios it is necessary to acquire a PI mutex if not
already locked by the current thread.

Provide a pi_mutex_lock_save() macro that tests if a mutex is already
locked by the current thread and if not locks it and saves the 'lock'
state. No error is returned if the lock is already owned by the
current thread.

The complementary pi_mutex_unlock_restore() macro unlocks the mutex if
the saved state was 'unlocked' (i.e. state == false). This will
restore the mutex to its previous state before calling
pi_mutex_lock_save().

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>